Tests passing
git-svn-id: svn+ssh://dev.real-comp.com/opt/usr/share/rcsvn/realcomp-data/branches/data-type-operations/realcomp-data@2824 cb9be8fc-bb24-4f4f-8dc9-9fc556ea96d8

+++ src/main/java/com/realcomp/data/DataType.java
@@ -113,7 +113,7 @@ public enum DataType {
         try {
             switch (DataType.getDataType(value)) {
                 case STRING:
-                    return Integer.parseInt(value.toString().isEmpty() ? "0" : value.toString());
+                    return ((Double) Double.parseDouble(value.toString().isEmpty() ? "0" : value.toString())).intValue();
                 case INTEGER:
                     return (Integer) value;
                 case FLOAT:
